+++ db/src/lib.rs
@@ -36,6 +36,7 @@ mod types;
 mod internal_types;
 mod upsert_resolution;
 mod values;
+mod tx;
 
 pub use types::DB;
